refactor(nodes-settings-dialog): seperate editor comp
