+++ dashboard/src/features/nodes/dialogs/index.ts
@@ -1,3 +1,3 @@
 export * from './mutation.dialog'
 export * from './delete-confirmation.dialog'
-export * from './settings.dialog'
+export * from './settings'
